create navigationBarBackground.swift
